backend/requirements.txt
asgiref==3.4.1
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==44.0.3
Deprecated==1.2.18
Django==3.2.25
django-cors-headers==3.10.1
djangorestframework==3.15.1
djangorestframework-simplejwt==4.3.0
img2pdf==0.6.1
importlib-resources==5.4.0
joblib==1.5.0
lxml==5.4.0
nltk==3.6.2
numpy==1.26.4
packaging==25.0
pandas==2.1.4
pdf2image==1.17.0
pdfminer.six==20250506
pikepdf==9.7.0
pillow==11.2.1
psycopg2==2.9.10
pycparser==2.22
PyJWT==1.7.1
PyPDF2==3.0.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-magic==0.4.27
pytz==2025.2
scikit-learn==1.3.2
scipy==1.13.1
six==1.17.0
sqlparse==0.4.4
threadpoolctl==3.6.0
typing_extensions==4.1.1
tzdata==2025.2
wrapt==1.17.2
zipp==3.6.0
